requirements.txt
torch==2.2.0
setuptools
wheel
torchvision
torchaudio
streamlit
sentence-transformers
audiocraft
music21
